+++ Test/Ordinaria/ContattiValidator.cs
@@ -1,5 +1,6 @@
 ﻿using FatturaElettronica.Ordinaria.FatturaElettronicaHeader.CedentePrestatore;
 using FluentValidation.TestHelper;
+using FluentValidation.Validators;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace FatturaElettronica.Test.Ordinaria
@@ -42,11 +43,16 @@ namespace FatturaElettronica.Test.Ordinaria
         public void EmailMustBeValid()
         {
             Challenge.Email = "not really";
-            Validator.ShouldHaveValidationErrorFor(x => x.Email, Challenge);
+            var result = Validator.TestValidate(Challenge);
+            result.ShouldHaveValidationErrorFor(x => x.Email);
+            
             Challenge.Email = "not@really";
-            Validator.ShouldHaveValidationErrorFor(x => x.Email, Challenge);
+            result = Validator.TestValidate(Challenge);
+            result.ShouldHaveValidationErrorFor(x => x.Email);
+            
             Challenge.Email = "maybe@we.can";
-            Validator.ShouldNotHaveValidationErrorFor(x => x.Email, Challenge);
+            result = Validator.TestValidate(Challenge);
+            result.ShouldNotHaveValidationErrorFor(x => x.Email);
         }
     }
 }