+++ Validators/DatiRiepilogoValidator.cs
@@ -11,7 +11,7 @@ namespace FatturaElettronica.Validators
         public DatiRiepilogoValidator()
         {
             RuleFor(x => x.Natura)
-                .SetValidator(new IsValidValidator<Natura>())
+                .SetValidator(new IsValidValidator<DatiRiepilogo, string, Natura>())
                 .When(x => !string.IsNullOrEmpty(x.Natura));
             RuleFor(x => x.Natura)
                 .Must(natura => !string.IsNullOrEmpty(natura))
@@ -33,7 +33,7 @@ namespace FatturaElettronica.Validators
                 .Must((challenge, _) => ImpostaValidateAgainstError00421(challenge))
                 .WithErrorCode("00421");
             RuleFor(x => x.EsigibilitaIVA)
-                .SetValidator(new IsValidValidator<EsigibilitaIVA>())
+                .SetValidator(new IsValidValidator<DatiRiepilogo, string, EsigibilitaIVA>())
                 .When(x => !string.IsNullOrEmpty(x.EsigibilitaIVA));
             RuleFor(x => x.RiferimentoNormativo)
                 .Length(1, 100)