+++ Validators/RegimeFiscaleValidator.cs
@@ -4,15 +4,15 @@ using FluentValidation.Validators;
 
 namespace FatturaElettronica.Validators
 {
-    public class RegimeFiscaleValidator<T> : IsValidValidator<T> where T : RegimeFiscale, new()
-    {
-        protected override ValidationFailure CreateValidationError(PropertyValidatorContext context)
-        {
-            var err = base.CreateValidationError(context);
-            if (context.PropertyValue != null && context.PropertyValue.ToString() == "RF03")
-                err.ErrorCode = "00459";
-
-            return err;
-        }
-    }
+    // public class RegimeFiscaleValidator<T> : IsValidValidator<T> where T : RegimeFiscale, new()
+    // {
+    //     protected override ValidationFailure CreateValidationError(PropertyValidatorContext context)
+    //     {
+    //         var err = base.CreateValidationError(context);
+    //         if (context.PropertyValue != null && context.PropertyValue.ToString() == "RF03")
+    //             err.ErrorCode = "00459";
+    //
+    //         return err;
+    //     }
+    // }
 }